requirements.txt
langchain
langchain-groq
python-dotenv
pinecone[grpc]
langchain-pinecone
pypdf
langchain-huggingface
langchain-community
flask